refactored minecraft service into two services, one for blocks and one for entities
