+++ src/main/proto/entities.proto
@@ -1,8 +1,12 @@
 syntax = "proto3";
 import "google/protobuf/empty.proto";
 import "google/protobuf/any.proto";
+import "sharedMessages.proto";
 package dk.itu.real.ooe;
 
+/**
+    The entity service
+*/
 service EntityService {
     /** Reads multiple entities **/
     rpc readEntities (Uuids) returns (Entities);
@@ -14,13 +18,6 @@ service EntityService {
     rpc readEntitiesInSphere(Sphere) returns (Entities);
 
     rpc updateEntityAI(EntityAIUpdate) returns (google.protobuf.Empty);
-}
-
-/** A point in 3D space */
-message Point {
-    int32 x = 1;
-    int32 y = 2;
-    int32 z = 3;
 }
 
 message Uuids {
